pyproject.toml
[project]
name = "google-agents"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.17.0",
    "google-generativeai>=0.8.5",
    "langfuse>=3.8.1",
    "litellm>=1.79.0",
    "openinference-instrumentation-google-adk>=0.1.6",
    "psutil>=7.1.2",
    "python-dotenv>=1.2.1",
    "yfinance>=0.2.66",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]
